ingestion-function/requirements.txt
functions-framework==3.*
google-cloud-vision==3.*
google-cloud-firestore==2.*
google-cloud-aiplatform>=1.38
google-cloud-storage==2.*